qa updates
