+++ src/lib/showMarketplace.ts
@@ -1,4 +1,4 @@
-import { Options } from "../types";
+import { Options } from "../types/options";
 import { assertInit } from "../utils/assertInit";
 import { setIframe } from "../utils/iframe";
