+++ src/types/global.d.ts
@@ -1,0 +1,2 @@
+// global variable from Webpack definePlugin
+declare const EMBEDDED_VERSION: string;
