--- src/types/index.ts
@@ -1,6 +1,0 @@
-export { Filters } from "./filters";
-export { Options } from "./options";
-export { Phrases } from "./phrases";
-export { ScreenConfiguration } from "./screenConfiguration";
-export { Theme } from "./theme";
-export { Translation } from "./translation";
