--- src/version.ts
@@ -1,1 +1,0 @@
-export const EMBEDDED_VERSION = "1.0.0";
