+++ webpack.config.js
@@ -1,4 +1,6 @@
 const path = require("path");
+const webpack = require("webpack");
+const package = require("./package.json");
 
 module.exports = {
   mode: "production",
@@ -29,4 +31,9 @@ module.exports = {
     library: "prismatic",
     libraryTarget: "umd",
   },
+  plugins: [
+    new webpack.DefinePlugin({
+      EMBEDDED_VERSION: JSON.stringify(package.version),
+    }),
+  ],
 };
